+++ pyquerytracker/__init__.py
@@ -1,6 +1,6 @@
 from .core import TrackQuery
 from .config import configure, ExportType
-from .export import export_data, get_summary, clear_data
+from .config import export_data, get_summary, clear_data
 
 __all__ = [
     "TrackQuery",
